requirements.txt
Django==3.1
Pillow==7.2.0
djangorestframework==3.11.1
whitenoise==5.2.0
six==1.15.0
django-cors-headers==3.4.0
gunicorn==20.0.4
django-filter==2.3.0
